Expose color for external settings icons
Marhsmallow now allows for system apps to add settings.  An icon is
loaded from the external package and a tint is applied to the icon
so that it matches other icons in the dashboard.  The tint color
uses the colorAccent attribute from the currently applied style.
Themes can overlay the colorAccent but this color is also used for
other elements of the UI which may not be ideal.

This patch adds a new color external_tile_icon_tint_color which
points to the attribute ?android:attr/colorAccent when not themed.
This retains the original look but offers better flexibility for
theme designers.

- Ported up to 7.1.1

Change-Id: Ifb4a99a315271cdcf978a796e3dea0eaef8d7750
(cherry picked from commit ac8bd9735aca0b1b0680a790fd24699e15ebe9be)